+++ src/messages/models/Template/mutations/pala_multi_gtz.js
@@ -43,10 +43,11 @@ export default {
             {
                 logo: Logo.PRIMARY.WHITE,
                 headline: { tag: 'small', br: ['low as'] },
+                subHeadline: 'small',
                 disclaimer: 'xsmall'
             }
         ],
-        ['ratio:1x4', { subHeadline: 'small' }],
+        ['ratio:1x4', { subHeadline: { tag: 'small', br: ['money'] } }],
         ['color:gray', { logo: Logo.PRIMARY.COLOR }],
         ['color:white', { logo: Logo.PRIMARY.COLOR }],
         ['color:white-no-border', { logo: Logo.PRIMARY.COLOR }]